Updated setMultiState Function
Updated the function that issues a setState for devices; this will ensure values already true in the device dps is not sent again in another command. While this will still send separate commands it will eliminate continously getting a light to turn ON while changing brightness in accordance with Issue #252 - this issue is now closed.